requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
3to2=1.1.1=pypi_0
appdirs=1.4.4=pypi_0
applib=1.2=pypi_0
blas=1.0=openblas
ca-certificates=2020.12.8=hecd8cb5_0
cachetools=4.2.0=pypi_0
certifi=2020.12.5=py39hecd8cb5_0
chardet=4.0.0=pypi_0
cmdln=2.0.0=pypi_0
esprima=4.0.1=pypi_0
google-api-core=1.24.1=pypi_0
google-api-python-client=1.12.8=pypi_0
google-auth=1.24.0=pypi_0
google-auth-httplib2=0.0.4=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
googleapis-common-protos=1.52.0=pypi_0
httplib2=0.18.1=pypi_0
idna=2.10=pypi_0
inflector=3.0.1=pypi_0
libclang-py3=3.9.0=pypi_0
libcxx=10.0.0=1
libedit=3.1.20191231=h1de35cc_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libopenblas=0.3.10=h0794777_0
ncurses=6.2=h0a44026_1
numpy=1.19.4=pypi_0
numpy-base=1.19.2=py39h3a452eb_0
oauthlib=3.1.0=pypi_0
openssl=1.1.1i=h9ed2024_0
ordered-set=4.0.2=pypi_0
pandas=1.1.5=pypi_0
pip=20.3.3=py39hecd8cb5_0
progress=1.5=pypi_0
protobuf=3.14.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
python=3.9.1=h88f2d9e_2
python-dateutil=2.8.1=pypi_0
pytz=2020.4=pypi_0
readline=8.0=h1de35cc_0
requests=2.25.1=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
setuptools=51.0.0=py39hecd8cb5_2
six=1.15.0=pypi_0
sqlite=3.33.0=hffcf06c_0
tk=8.6.10=hb0a8c7a_0
tzdata=2020d=h14c3975_0
uritemplate=3.0.1=pypi_0
urllib3=1.26.2=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
xz=5.2.5=h1de35cc_0
zlib=1.2.11=h1de35cc_3
zope-cachedescriptors=4.3.1=pypi_0
requests~=2.25.1